+++ src/main/java/com/github/lipinskipawel/board/ai/ml/Func.java
@@ -1,8 +1,0 @@
-
-package com.github.lipinskipawel.board.ai.ml;
-
-
-public interface Func {
-
-    double apply(double val);
-}