+++ src/main/java/com/github/lipinskipawel/board/ai/ml/lossfunction/LossFunction.java
@@ -1,9 +1,0 @@
-package com.github.lipinskipawel.board.ai.ml.lossfunction;
-
-import com.github.lipinskipawel.board.ai.ml.Matrix;
-
-abstract public class LossFunction {
-
-
-    abstract public Matrix compute(Matrix target, Matrix output);
-}