+++ src/main/java/module-info.java
@@ -7,6 +7,4 @@ module game.engine {
 
     opens com.github.lipinskipawel.board.engine;
     opens com.github.lipinskipawel.board.ai.bruteforce;
-    opens com.github.lipinskipawel.board.ai.ml;
-    opens com.github.lipinskipawel.board.ai.ml.activation;
 }